backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-multipart
pydantic-settings
transformers
torch
pillow
numpy
opencv-python
